Split both schemas in different files and URLs
